Use regexp_instr explicitly on Redshift (#102)
with the change to adapter resolution in dbt 0.20, redshift has been
using the postgres version of the macro, which fails because the
function `regexp_matches` exists on Postgres but not Redshift. While the
`regexp_instr` does exist on Redshift, so the default was correct all
along.

This commit simply adds an explicit adapter for Redshift that is the
same as the default adapter.