Cast `relation_column` and `relation_column_type` to string type
In some circumstances, Redshift will fail to execute queries generated by `expect_column_values_to_be_in_type_list` if it is unable infer the type of the `relation_column` or `relation_column_type` fields. This commit enforces a casting of those values to a string type to prevent ambiguity.